refactor: split tree shake lazy barrel tests
